ml/catcheese/requirements.txt
backports.functools-lru-cache==1.4
click==6.7
cycler==0.10.0
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
matplotlib==2.1.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
six==1.11.0
subprocess32==3.2.7
torch==0.2.0.post3
torchvision==0.1.9
Werkzeug==0.12.2
